+++ src/main/java/com/bupt/uta/Service/ProductService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface ProductService{
     R<List<Product>> getAllProduct();
-    R<Product> getProduct(int id);
-    R<List<Product>> getSuppProduct(int supplierid);
+    R<Product> getProduct(long id);
+    R<List<Product>> getSuppProduct(long supplierid);
     R<String> addProduct(Product product);
-    R<String> deleteProduct(int id);
+    R<String> deleteProduct(long id);
     R<String> updateProduct(Product product);
 }
